usqyr-ai/requirements.txt
Django==5.2
django-htmlmin==0.11.0
django-ipware==7.0.1
django-jazzmin==3.0.1
djangorestframework==3.16.1
psycopg2-binary==2.9.11
openai==2.8.1
httpx==0.28.1
pyTelegramBotAPI==4.29.1
python-decouple==3.8
python-ipware==3.0.0
tqdm==4.67.1

